エンコード手動追加を /api/recorded/{id}/encode へ変更
